+++ packages/icons/src/action/index.ts
@@ -7,6 +7,8 @@ export {
   Delete,
   Copy,
   Refresh,
+  ThumbUp,
+  ThumbDown,
   Upload,
   CloudUpload,
   EyeOpen,